VM bootstrap with working P=V and jumping into the KVA tables :)

+++ kernel/lib/stddef.h
@@ -5,5 +5,5 @@
 #define offsetof(TYPE, MEMBER) ((size_t) & ((TYPE*)0)->MEMBER)
 typedef __PTRDIFF_TYPE__ ptrdiff_t;
 typedef __SIZE_TYPE__ size_t;
-typedef unsigned long uintptr_t;
+typedef unsigned long long uintptr_t;
 #endif /* LIB_STDDEF */
